Update onComplete functionality
Remove all .to() and last_run_tween items to a cleaner approach.
OnComplete is added to the first tween of every animation set
and once that set completes whatever is next in the queue of
droplet JSON will be animated. Also add notification once
the server is closed to prevent persistent calls.