Add test for process isolation failure
The creation of multiple processes appears to work, however
operations using those processes do not. Using crossbeam to
create a scoped thread in order to guarantee the manager will
not go away before a thread completes. This may not be necessary
as get_new_process was updated to no longer expect a mutable manager.